Remove LineNr, Todo background color
This is something I can never make up my mind about. On the one hand, it
is nice to have a clear delineation between the buffer's contents and
the line numbers, but there just isn't enough padding around the line
numbers to make it look neat.

The previous coloration of Todo made it invisible on a line with
CursorLine highlighting.

Additionally, link SignColumn to LineNr.